optional autosave
fixes #389

+++ src/js/window/main-window.js
@@ -1067,7 +1067,7 @@ let markBoardFileDirty = () => {
   boardFileDirtyTimer = setTimeout(saveBoardFile, 5000)
 }
 
-let saveBoardFile = () => {
+let saveBoardFile = (opt = { force: false }) => {
   // are we still drawing?
   if (storyboarderSketchPane.getIsDrawingOrStabilizing()) {
     // wait, then retry
@@ -1075,13 +1075,14 @@ let saveBoardFile = () => {
     return
   }
 
-
   if (boardFileDirty) {
     clearTimeout(boardFileDirtyTimer)
     boardData.version = pkg.version
-    fs.writeFileSync(boardFilename, JSON.stringify(boardData, null, 2))
-    boardFileDirty = false
-    console.log('saved board file:', boardFilename)
+    if (opt.force || prefsModule.getPrefs()['enableAutoSave']) {
+      fs.writeFileSync(boardFilename, JSON.stringify(boardData, null, 2))
+      boardFileDirty = false
+      console.log('saved board file:', boardFilename)
+    }
   }
 }
 
@@ -3219,10 +3220,15 @@ const exportCleanup = () => {
 
 let save = () => {
   saveImageFile()
-  saveBoardFile()
+  saveBoardFile({ force: true })
   sfx.positive()
-  notifications.notify({message: "Saving is automatic. I already saved before you pressed this, so you don't really need to save at all. \n\nBut I did want you to know, that I think you're special - and I like you just the way you are.\n\nHere's a story tip..." , timing: 15})
-  setTimeout(()=>{storyTips.show()}, 1000)
+
+  if (prefsModule.getPrefs()['enableAutoSave']) {
+    notifications.notify({message: "Saving is automatic. I already saved before you pressed this, so you don't really need to save at all. \n\nBut I did want you to know, that I think you're special - and I like you just the way you are.\n\nHere's a story tip..." , timing: 15})
+    setTimeout(()=>{storyTips.show()}, 1000)
+  } else {
+    notifications.notify({ message: "Saved." })
+  }
 }
